QmlDesigner: Add progress bar and time estimate to light baking dialog
In Qt 6.10 we get progress and estimated time remaining information
from the baking process, so we can now show a meaningful progress bar
and time information on baking dialog.

Fixes: QDS-16099

+++ src/plugins/qmldesigner/components/edit3d/bakelights.cpp
@@ -186,7 +186,7 @@ void BakeLights::bakeLights()
             if (prog >= 0)
                 emit progressChanged(prog);
             if (timeRemaining >= 0)
-                emit timeRemainingChanged(timeRemaining / 1000.);
+                emit timeRemainingChanged(int(timeRemaining / 1000.));
         };
     } else {
         progressCallback = [this](const QVariant &data) {
